Add PhiloxStateless
